Group files for merge to final
