+++ bitterdispute/elementary_functions.py
@@ -9,11 +9,11 @@ def sin(x):
         new_x = Variable(name=x.name, value=np.sin(x.val), derivative=x.der)
 
         for key in x.der:
-            new_x.der[key] = x.der[key]*np.cos(x.val)
+            new_x.der[key] = x.der.get(key)*np.cos(x.val)
         return new_x
     except AttributeError: # constant
         return np.sin(x)
-    
+
 
 def cos(x): #--> -sin
     """Defines what happens when cosine operations performed on
@@ -22,7 +22,7 @@ def cos(x): #--> -sin
     try:
         new_x = Variable(name = x.name, value = np.cos(x.val), derivative = x.der)
         for key in x.der:
-            x.der[key] = x.der[key]*(-np.sin(x.val))
+            x.der[key] = x.der.get(key)*(-np.sin(x.val))
         return new_x
     except AttributeError:
         return np.cos(x)
@@ -35,12 +35,12 @@ def tan(x): #--> 1/cos^2(x)
     try:
         new_x = Variable(name = x.name, value = np.tan(x.val), derivative = x.der)
         for key in x.der:
-            x.der[key] = x.der[key]*(1 / (np.cos(x.val) ** 2))
+            x.der[key] = x.der.get(key)*(1 / (np.cos(x.val) ** 2))
         return new_x
     except AttributeError:
         return np.tan(x)
 
-                                    
+
 def sqrt(x):
     """Defines what happens when square root operations performed on
     a Variable() object or a constant value. Includes calculation of derivative.
@@ -51,11 +51,11 @@ def sqrt(x):
             raise ValueError('Cannot evaluate the square root of a negative number')
         x_new = Variable(name = x.name, value = np.sqrt(x.val), derivative = x.der)
         for key in x.der:
-            x.der[key] = x.der[key]*(1/(2*np.sqrt(x.val)))
+            x.der[key] = x.der.get(key)*(1/(2*np.sqrt(x.val)))
         return x_new
     except AttributeError:
         return np.sqrt(x)
-                                    
+
 
 def log(x, base=math.e):
     """Defines what happens when log operations performed on
@@ -64,22 +64,22 @@ def log(x, base=math.e):
     try:
         #Ensure input domain valid
         if x.val <=0:
-            raise ValueError('Cannot evaluate the log of a non-positive number')       
+            raise ValueError('Cannot evaluate the log of a non-positive number')
         new_x = Variable(name = x.name, value = math.log(x.val, base), derivative = x.der)
         for key in x.der:
-            x.der[key] = x.der[key]/(x.val * math.log(base))
+            x.der[key] = x.der.get(key)/(x.val * math.log(base))
         return new_x
     except AttributeError:
         return math.log(x, base)
-                                     
-                                     
+
+
 def exp(x):
     '''Returns Variable obj of exp(x)
     '''
     try:
-        new_x = Variable(name=x.name, value = np.exp(x.val), derivative = x.der)       
+        new_x = Variable(name=x.name, value = np.exp(x.val), derivative = x.der)
         for key in x.der:
-            new_x.der[key] = x.der[key]*new_x.val      
+            new_x.der[key] = x.der.get(key)*new_x.val
         return new_x
     except AttributeError: # constant
         return np.exp(x)
@@ -91,27 +91,27 @@ def arcsin(x):
     try:
         if x.val < -1.0 or x.val > 1.0:
             raise ValueError("input of arcsin should within (-1, 1)")
-        
-        new_x = Variable(name=x.name, value = np.arcsin(x.val), derivative = x.der)       
+
+        new_x = Variable(name=x.name, value = np.arcsin(x.val), derivative = x.der)
         for key in x.der:
-            new_x.der[key] = 1/np.sqrt(1 - x.val**2)*x.der[key]
+            new_x.der[key] = 1/np.sqrt(1 - x.val**2)*x.der.get(key)
         return new_x
     except AttributeError: # constant
         if x < -1.0 or x > 1.0:
             raise ValueError("input of arcsin should within (-1, 1)")
         return np.arcsin(x)
 
-    
+
 def arccos(x):
     """Returns Variable obj ofarccos(x)
     """
     try:
         if x.val < -1.0 or x.val > 1.0:
             raise ValueError("input of arccos should within (-1, 1)")
-            
+
         new_x = Variable(name=x.name, value = np.arccos(x.val), derivative = x.der)
         for key in x.der:
-            new_x.der[key] = -1/np.sqrt(1 - x.val**2)*x.der[key]
+            new_x.der[key] = -1/np.sqrt(1 - x.val**2)*x.der.get(key)
         return new_x
     except AttributeError: # constant
         if x < -1.0 or x > 1.0:
@@ -125,7 +125,7 @@ def arctan(x):
     try:
         new_x = Variable(name=x.name, value = np.arctan(x.val), derivative = x.der)
         for key in x.der:
-            new_x.der[key] = 1/(1+x.val**2)*x.der[key]
+            new_x.der[key] = 1/(1+x.val**2)*x.der.get(key)
         return new_x
     except AttributeError: # constant
         return np.arctan(x)
@@ -135,10 +135,10 @@ def sinh(x):
     a Variable() object or a constant value. Includes calculation of derivative.
     '''
     try:
-        # Create a Variable instance 
+        # Create a Variable instance
         new_x = Variable(name=x.name, value = np.sinh(x.val), derivative = x.der)
         for key in x.der:
-            new_x.der[key] = x.der[key]*np.cosh(x.val)
+            new_x.der[key] = x.der.get(key)*np.cosh(x.val)
         return new_x
     except AttributeError: # if x = constant
         return math.sinh(x)
@@ -149,16 +149,16 @@ def cosh(x):
     a Variable() object or a constant value. Includes calculation of derivative.
     '''
     try:
-        # Create a Variable instance 
+        # Create a Variable instance
         new_x = Variable(name=x.name, value = np.cosh(x.val), derivative = x.der)
         for key in x.der:
             # quality check
-            new_x.der[key] = x.der[key]*np.sinh(x.val)
+            new_x.der[key] = x.der.get(key)*np.sinh(x.val)
         return new_x
     except AttributeError: # if x = constant
         return math.cosh(x)
-    
-    
+
+
 def tanh(x):
     '''Defines what happens when hyperbolic tangent operations performed on
     a Variable() object or a constant value. Includes calculation of derivative.
@@ -169,7 +169,7 @@ def tanh(x):
         new_x = Variable(name=x.name, value = np.tanh(x.val), derivative = x.der)
         for key in x.der:
             # quality check
-            new_x.der[key] = x.der[key]*((1.0/np.cosh(x.val))**2)
+            new_x.der[key] = x.der.get(key)*((1.0/np.cosh(x.val))**2)
         return new_x
     except AttributeError: # if x = constant
         return math.tanh(x)
