+++ bitterdispute/optimization.py
@@ -1,0 +1,2 @@
+from optimization_gradientdescent_linesearch import *
+from optimization_newton_quasi_newton import *
