+++ tests/test_forward_mode.py
@@ -17,7 +17,7 @@ def test_standard_formula1(mocker, capsys):
     """
     mocker.patch('builtins.input', side_effect=[2, 3, 5, '4*x+7*y'])
     x = AD()
-    assert x.values == [41]
+    assert x.outputs == [41]
     assert x.derivatives == [{'x': 4.0, 'y': 7.0}]
     assert x.formulas == ['4*x+7*y']
 
@@ -32,7 +32,7 @@ def test_standard_formula2(mocker):
     """
     mocker.patch('builtins.input', side_effect=[3, 3, 5, 8, '4/x-7*y+z/2'])
     x = AD()
-    assert x.values == [-33]
+    assert x.outputs == [-33]
     assert x.derivatives == [{'z': 0.5, 'x': -0.0625, 'y': -7.0}]
     assert x.formulas == ['4/x-7*y+z/2']
 
@@ -46,7 +46,7 @@ def test_elementary_functions1(mocker):
     """
     mocker.patch('builtins.input', side_effect=[2, 5, 8, 'sin(x)*7+4*tanh(y)'])
     x = AD()
-    assert [round(val, 10) for val in x.values] == [round(10.925144543414053, 10)]
+    assert [round(val, 10) for val in x.outputs] == [round(10.925144543414053, 10)]
     assert x.derivatives == [{'y': 0.0007263329237752267, 'x': -1.0185002366602949}]
     assert x.formulas == ['sin(x)*7+4*tanh(y)']
 
@@ -61,7 +61,7 @@ def test_elementary_functions2(mocker):
     """
     mocker.patch('builtins.input', side_effect=[3, 2, 7, 1, 'sin(cos(tan(x)*y)*z)'])
     x = AD()
-    assert [round(val, 10) for val in x.values] == [round(-0.18607529153476626, 10)]
+    assert [round(val, 10) for val in x.outputs] == [round(-0.18607529153476626, 10)]
     assert x.derivatives == [{'y': 3.046986001124803, 'z': -0.09194869423196683, 'x': 46.91292723071239}]
     assert x.formulas == ['sin(cos(tan(x)*y)*z)']
 
@@ -74,7 +74,7 @@ def test_power(mocker):
     """
     mocker.patch('builtins.input', side_effect=[1, 3, 'x**5'])
     x = AD()
-    assert x.values == [243]
+    assert x.outputs == [243]
     assert x.derivatives == [{'x': 405.0}]
     assert x.formulas == ['x**5']
 
@@ -88,7 +88,7 @@ def test_positive(mocker):
     """
     mocker.patch('builtins.input', side_effect=[2, 2, 8, '-x+3*(y)'])
     x = AD()
-    assert x.values == [-2]
+    assert x.outputs == [-2]
     assert x.derivatives == [{'x': -1, 'y': 3.0}]
     assert x.formulas == ['-x+3*(y)']
